src/requirements.txt
pandas==1.1.3
xgboost==2.0.3
python-dateutil==2.8.1
joblib
flask
gunicorn
pyarrow
python-dotenv